webserver.catalog settings
